% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.dx.R
\name{set.dx}
\alias{set.dx}
\title{set.dx}
\usage{
set.dx(K = 47, S = 1)
}
\arguments{
\item{K}{parameter for dx generator}

\item{S}{parameter for dx generator}
}
\value{
Creates an initialized table of size k
}
\description{
set.dx
}
\examples{
set.dx(K=47, S=1)
}
